# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookify API Documentation",
        "description": "API documentation for Bookify application",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/v1/payments": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a payment request for an order",
                "operationId": "ea58c183f7c6e4817c3a6700f9b2e4e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "gateway"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "string",
                                        "example": "order-23091"
                                    },
                                    "gateway": {
                                        "type": "string",
                                        "example": "midtrans"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment created successfully"
                    },
                    "500": {
                        "description": "Failed to create payment"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/payments/webhook/midtrans": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle Midtrans webhook",
                "operationId": "528b900ecd5bb4dba0295dfa4bd1fb95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "transaction_status": "settlement",
                                "order_id": "order-23091"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully"
                    },
                    "500": {
                        "description": "Webhook processing error"
                    }
                }
            }
        },
        "api/v1/payments/webhook/xendit": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle Xendit webhook",
                "operationId": "55ab2aa55940050e1c9c813e2d0720be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "status": "PAID",
                                "order_id": "order-23091"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully"
                    },
                    "500": {
                        "description": "Webhook processing error"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}